+++ src/actions/remove.ts
@@ -1,9 +1,0 @@
-import { remove } from '../core/fileTransferTask';
-import createFileAction from './createFileAction';
-
-export const removeRemote = createFileAction('remove', (source, config, { remoteFs }) =>
-  remove(config.remotePath, remoteFs, {
-    ignore: config.ignore,
-    skipDir: config.skipDir,
-  })
-);