+++ src/extension.ts
@@ -2,18 +2,20 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import sftpBarItem from './ui/sftpBarItem';
-import initCommand from './commands/init';
-import reportError from './helper/reportError';
+import app from './app';
+import initCommands from './initCommands';
+import { reportError } from './helper';
 import fileActivityMonitor from './modules/fileActivityMonitor';
-import { initConfigs } from './modules/config';
-import { endAllRemote } from './modules/remoteFs';
-import appState from './modules/appState';
-import { watchFiles, clearAllWatcher } from './modules/fileWatcher';
+import { tryLoadConfigs } from './modules/config';
+import { getAllFileService, createFileService, disposeFileService } from './modules/serviceManager';
 import { getWorkspaceFolders, setContextValue } from './host';
+import RemoteExplorer from './modules/remoteExplorer';
 
-function setupWorkspaceFolder(dir) {
-  return initConfigs(dir).then(watchFiles);
+async function setupWorkspaceFolder(dir) {
+  const configs = await tryLoadConfigs(dir);
+  configs.forEach(config => {
+    createFileService(config, dir);
+  });
 }
 
 function setup(workspaceFolders: vscode.WorkspaceFolder[]) {
@@ -25,8 +27,12 @@ function setup(workspaceFolders: vscode.WorkspaceFolder[]) {
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
-export function activate(context: vscode.ExtensionContext) {
-  initCommand(context);
+export async function activate(context: vscode.ExtensionContext) {
+  try {
+    initCommands(context);
+  } catch (error) {
+    reportError(error, 'initCommands');
+  }
 
   const workspaceFolders = getWorkspaceFolders();
   if (!workspaceFolders) {
@@ -34,19 +40,26 @@ export function activate(context: vscode.ExtensionContext) {
   }
 
   setContextValue('enabled', true);
-  sftpBarItem.show();
-
-  appState.subscribe(state => {
-    const currentText = sftpBarItem.getText();
-    if (currentText.endsWith('SFTP')) {
-      sftpBarItem.reset();
+  app.sftpBarItem.show();
+  app.state.subscribe(_ => {
+    const currentText = app.sftpBarItem.getText();
+    // current is showing profile
+    if (currentText.startsWith('SFTP')) {
+      app.sftpBarItem.reset();
+    }
+    if (app.remoteExplorer) {
+      app.remoteExplorer.refresh();
     }
   });
-  setup(workspaceFolders).catch(reportError);
+  try {
+    await setup(workspaceFolders);
+    app.remoteExplorer = new RemoteExplorer(context);
+  } catch (error) {
+    reportError(error);
+  }
 }
 
 export function deactivate() {
-  clearAllWatcher();
-  endAllRemote();
   fileActivityMonitor.destory();
+  getAllFileService().forEach(disposeFileService);
 }