+++ src/helper/select.ts
@@ -1,56 +1,64 @@
 import * as vscode from 'vscode';
-import FileSystem, { FileType } from '../core/Fs/FileSystem';
-import { getAllRawConfigs } from '../modules/config';
+import { FileSystem, FileType } from '../core';
 import * as path from 'path';
 
 const ROOT = '@root';
 
-interface IFileLookUp {
-  [x: string]: IFilePickerItem[];
+interface IFileLookUp<T> {
+  [x: string]: T[];
 }
 
 interface IFilePickerOption {
   type?: FileType;
-  filter?: (file: IFilePickerItem) => boolean;
 }
 
-interface IFilePickerItem {
+interface FileListChildItem extends FileListItem {
+  parentFsPath: string;
+}
+
+interface FileListItem {
   name: string;
   fsPath: string;
-  parentFsPath: string;
   type: FileType;
+
   description: string;
 
-  // config index
-  index: number;
-  getFs?: () => Promise<FileSystem>;
+  getFs: (() => Promise<FileSystem>) | FileSystem;
+  filter?: (x: string) => boolean;
 }
 
-async function showFiles(
-  fileLookUp: IFileLookUp,
-  parent: IFilePickerItem | null,
-  files: IFilePickerItem[],
+async function showFiles<T extends FileListChildItem>(
+  fileLookUp: IFileLookUp<T>,
+  parent: T | null,
+  files: T[],
   option: IFilePickerOption = {}
-) {
-  let avalibleFiles = files.slice();
+): Promise<T | undefined> {
+  let avalibleFiles = files;
   let filter;
-  let fileFilter;
   if (option.type === FileType.Directory) {
-    fileFilter = file => file.type === FileType.Directory;
+    filter = file => file.type === FileType.Directory;
   } else {
     // don't show SymbolicLink
-    fileFilter = file => file.type !== FileType.SymbolicLink;
+    filter = file => file.type !== FileType.SymbolicLink;
   }
-
-  if (option.filter !== undefined) {
+  if (parent && parent.filter) {
+    const parentFilter = parent.filter;
+    const oldFilter = filter;
     filter = file => {
-      return fileFilter(file) && option.filter(file);
+      return oldFilter(file) && parentFilter(file);
     };
-  } else {
-    filter = fileFilter;
   }
+  avalibleFiles = avalibleFiles.filter(file => {
+    if (file.parentFsPath === ROOT) {
+      return true;
+    }
 
-  avalibleFiles = avalibleFiles.filter(filter);
+    if (file.name === '.' || file.name === '..') {
+      return true;
+    }
+
+    return filter(file);
+  });
 
   const items = avalibleFiles
     .map(file => ({
@@ -93,8 +101,8 @@ async function showFiles(
 
   const selectedValue = result.value;
   const selectedPath = selectedValue.fsPath;
-  // fs will be nerver be null if current is root, so get fs from picker item
-  const fileSystem = await selectedValue.getFs();
+  const fileSystem =
+    typeof selectedValue.getFs === 'function' ? await selectedValue.getFs() : selectedValue.getFs;
 
   const nextItems = fileLookUp[selectedPath];
   if (nextItems !== undefined) {
@@ -102,36 +110,36 @@ async function showFiles(
   }
 
   return fileSystem.list(selectedPath).then(subFiles => {
-    const subItems = subFiles.map(file => ({
-      name: path.basename(file.fspath) + (file.type === FileType.Directory ? '/' : ''),
-      fsPath: file.fspath,
-      parentFsPath: selectedPath,
-      type: file.type,
-      description: '',
-      getFs: selectedValue.getFs,
-      index: selectedValue.index,
-    }));
+    const subItems = subFiles.map(file =>
+      Object.assign({}, selectedValue, {
+        name: path.basename(file.fspath) + (file.type === FileType.Directory ? '/' : ''),
+        fsPath: file.fspath,
+        parentFsPath: selectedPath,
+        type: file.type,
+        description: '',
+      })
+    );
 
-    subItems.unshift({
-      name: '..',
-      fsPath: selectedValue.parentFsPath,
-      parentFsPath: '#will never reach here, cause the dir has alreay be cached#',
-      type: FileType.Directory,
-      description: 'go back',
-      getFs: selectedValue.getFs,
-      index: selectedValue.index,
-    });
+    subItems.unshift(
+      Object.assign({}, selectedValue, {
+        name: '..',
+        fsPath: selectedValue.parentFsPath,
+        parentFsPath: '#will never reach here, cause the dir has alreay be cached#',
+        type: FileType.Directory,
+        description: 'go back',
+      })
+    );
 
     if (allowChooseFolder) {
-      subItems.unshift({
-        name: '.',
-        fsPath: selectedPath,
-        parentFsPath: selectedValue.parentFsPath,
-        type: FileType.Directory,
-        description: 'choose current folder',
-        getFs: selectedValue.getFs,
-        index: selectedValue.index,
-      });
+      subItems.unshift(
+        Object.assign({}, selectedValue, {
+          name: '.',
+          fsPath: selectedPath,
+          parentFsPath: selectedValue.parentFsPath,
+          type: FileType.Directory,
+          description: 'choose current folder',
+        })
+      );
     }
 
     fileLookUp[selectedPath] = subItems;
@@ -139,59 +147,13 @@ async function showFiles(
   });
 }
 
-export function listFiles(
-  items: Array<{
-    name?: string;
-    description: string;
-    fsPath: string;
-    getFs?: () => Promise<FileSystem>;
-    index: number;
-  }>,
-  option: IFilePickerOption
-) {
-  const baseItems = items.map(item => ({
-    name: item.name || item.fsPath,
-    fsPath: item.fsPath,
-    parentFsPath: ROOT,
-    type: FileType.Directory,
-    description: item.description,
-    getFs: item.getFs,
-    index: item.index,
-  }));
+export function listFiles<T extends FileListItem>(
+  items: T[],
+  option?: IFilePickerOption
+): Promise<T & FileListChildItem | undefined> {
+  const baseItems = items.map(item => Object.assign({}, item, { parentFsPath: ROOT }));
   const fileLookUp = {
     [ROOT]: baseItems,
   };
   return showFiles(fileLookUp, null, baseItems, option);
-}
-
-export function selectContext(): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const configs = getAllRawConfigs();
-    const projectsList = configs
-      .map(cfg => ({
-        value: cfg.context,
-        label: cfg.name || vscode.workspace.asRelativePath(cfg.context),
-        description: '',
-        detail: cfg.context,
-      }))
-      .sort((l, r) => l.label.localeCompare(r.label));
-
-    // if (projectsList.length === 1) {
-      // return resolve(projectsList[0].value);
-    // }
-
-    vscode.window
-      .showQuickPick(projectsList, {
-        placeHolder: 'Select a folder...',
-      })
-      .then(selection => {
-        if (selection) {
-          resolve(selection.value);
-          return;
-        }
-
-        // cancel selection
-        return null;
-      }, reject);
-  });
 }