+++ src/modules/appState.ts
@@ -1,12 +1,16 @@
 class AppState {
-  private _profile: string;
+  private _profile: string | null = null;
   private _observer: (x: any) => void;
 
-  get profile(): string {
+  get profile(): string | null {
     return this._profile;
   }
 
-  set profile(newProfile: string) {
+  set profile(newProfile: string | null) {
+    if (this._profile === newProfile) {
+      return;
+    }
+
     this._profile = newProfile;
     this._observer(this.getStateSnapshot());
   }
@@ -22,6 +26,4 @@ class AppState {
   }
 }
 
-const appState = new AppState();
-
-export default appState;
+export default AppState;