+++ src/modules/localFs.ts
@@ -1,6 +1,0 @@
-import * as path from 'path';
-import LocalFileSystem from '../core/Fs/LocalFileSystem';
-
-const fs = new LocalFileSystem(path);
-
-export default fs;